+++ libraries/transformer/src/main/java/androidx/media3/transformer/DefaultAssetLoaderFactory.java
@@ -16,114 +16,81 @@
 
 package androidx.media3.transformer;
 
-import static androidx.media3.common.util.Assertions.checkStateNotNull;
-
 import android.content.Context;
 import android.os.Looper;
-import androidx.annotation.Nullable;
 import androidx.media3.common.MediaItem;
 import androidx.media3.common.util.Clock;
 import androidx.media3.common.util.UnstableApi;
-import androidx.media3.exoplayer.source.MediaSource;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 
 /** The default {@link AssetLoader.Factory} implementation. */
 @UnstableApi
 public final class DefaultAssetLoaderFactory implements AssetLoader.Factory {
 
-  @Nullable private Context context;
-  @Nullable private MediaItem mediaItem;
-  private boolean removeAudio;
-  private boolean removeVideo;
-  private boolean flattenVideoForSlowMotion;
-  @Nullable private MediaSource.Factory mediaSourceFactory;
-  @Nullable private Codec.DecoderFactory decoderFactory;
-  @Nullable private Looper looper;
-  @Nullable private AssetLoader.Listener listener;
-  @Nullable private Clock clock;
+  private final AssetLoader.Factory assetLoaderFactory;
+
+  /** Creates an instance. */
+  public DefaultAssetLoaderFactory() {
+    assetLoaderFactory = new ExoPlayerAssetLoader.Factory();
+  }
 
   @Override
   @CanIgnoreReturnValue
   public AssetLoader.Factory setContext(Context context) {
-    this.context = context;
-    return this;
+    return assetLoaderFactory.setContext(context);
   }
 
   @Override
   @CanIgnoreReturnValue
   public AssetLoader.Factory setMediaItem(MediaItem mediaItem) {
-    this.mediaItem = mediaItem;
-    return this;
+    return assetLoaderFactory.setMediaItem(mediaItem);
   }
 
   @Override
   @CanIgnoreReturnValue
   public AssetLoader.Factory setRemoveAudio(boolean removeAudio) {
-    this.removeAudio = removeAudio;
-    return this;
+    return assetLoaderFactory.setRemoveAudio(removeAudio);
   }
 
   @Override
   @CanIgnoreReturnValue
   public AssetLoader.Factory setRemoveVideo(boolean removeVideo) {
-    this.removeVideo = removeVideo;
-    return this;
+    return assetLoaderFactory.setRemoveVideo(removeVideo);
   }
 
   @Override
   @CanIgnoreReturnValue
   public AssetLoader.Factory setFlattenVideoForSlowMotion(boolean flattenVideoForSlowMotion) {
-    this.flattenVideoForSlowMotion = flattenVideoForSlowMotion;
-    return this;
-  }
-
-  @Override
-  @CanIgnoreReturnValue
-  public AssetLoader.Factory setMediaSourceFactory(MediaSource.Factory mediaSourceFactory) {
-    this.mediaSourceFactory = mediaSourceFactory;
+    assetLoaderFactory.setFlattenVideoForSlowMotion(flattenVideoForSlowMotion);
     return this;
   }
 
   @Override
   @CanIgnoreReturnValue
   public AssetLoader.Factory setDecoderFactory(Codec.DecoderFactory decoderFactory) {
-    this.decoderFactory = decoderFactory;
-    return this;
+    return assetLoaderFactory.setDecoderFactory(decoderFactory);
   }
 
   @Override
   @CanIgnoreReturnValue
   public AssetLoader.Factory setLooper(Looper looper) {
-    this.looper = looper;
-    return this;
+    return assetLoaderFactory.setLooper(looper);
   }
 
   @Override
   @CanIgnoreReturnValue
   public AssetLoader.Factory setListener(AssetLoader.Listener listener) {
-    this.listener = listener;
-    return this;
+    return assetLoaderFactory.setListener(listener);
   }
 
   @Override
   @CanIgnoreReturnValue
   public AssetLoader.Factory setClock(Clock clock) {
-    this.clock = clock;
-    return this;
+    return assetLoaderFactory.setClock(clock);
   }
 
   @Override
   public AssetLoader createAssetLoader() {
-    return new ExoPlayerAssetLoader(
-        checkStateNotNull(context),
-        checkStateNotNull(mediaItem),
-        removeAudio,
-        removeVideo,
-        flattenVideoForSlowMotion,
-        checkStateNotNull(mediaSourceFactory),
-        checkStateNotNull(decoderFactory),
-        checkStateNotNull(looper),
-        checkStateNotNull(listener),
-        checkStateNotNull(clock));
+    return assetLoaderFactory.createAssetLoader();
   }
 }