+++ libraries/transformer/src/main/java/androidx/media3/transformer/Transformer.java
@@ -43,10 +43,7 @@ import androidx.media3.effect.GlEffect;
 import androidx.media3.effect.GlEffectsFrameProcessor;
 import androidx.media3.effect.GlMatrixTransformation;
 import androidx.media3.exoplayer.audio.SonicAudioProcessor;
-import androidx.media3.exoplayer.source.DefaultMediaSourceFactory;
 import androidx.media3.exoplayer.source.MediaSource;
-import androidx.media3.extractor.DefaultExtractorsFactory;
-import androidx.media3.extractor.mp4.Mp4Extractor;
 import com.google.common.collect.ImmutableList;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import java.lang.annotation.Documented;
@@ -54,7 +51,6 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 import java.util.List;
-import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 
 /**
  * A transformer to transform media inputs.
@@ -91,7 +87,6 @@ public final class Transformer {
     private boolean removeVideo;
     private boolean forceSilentAudio;
     private ListenerSet<Transformer.Listener> listeners;
-    private MediaSource.@MonotonicNonNull Factory mediaSourceFactory;
     private AssetLoader.Factory assetLoaderFactory;
     private Codec.DecoderFactory decoderFactory;
     private Codec.EncoderFactory encoderFactory;
@@ -132,7 +127,6 @@ public final class Transformer {
       this.removeVideo = transformer.removeVideo;
       this.forceSilentAudio = transformer.forceSilentAudio;
       this.listeners = transformer.listeners;
-      this.mediaSourceFactory = transformer.mediaSourceFactory;
       this.assetLoaderFactory = transformer.assetLoaderFactory;
       this.decoderFactory = transformer.decoderFactory;
       this.encoderFactory = transformer.encoderFactory;
@@ -291,21 +285,6 @@ public final class Transformer {
     @CanIgnoreReturnValue
     public Builder removeAllListeners() {
       this.listeners.clear();
-      return this;
-    }
-
-    /**
-     * Sets the {@link MediaSource.Factory} to be used to retrieve the inputs to transform.
-     *
-     * <p>The default value is a {@link DefaultMediaSourceFactory} built with the context provided
-     * in {@linkplain #Builder(Context) the constructor}.
-     *
-     * @param mediaSourceFactory A {@link MediaSource.Factory}.
-     * @return This builder.
-     */
-    @CanIgnoreReturnValue
-    public Builder setMediaSourceFactory(MediaSource.Factory mediaSourceFactory) {
-      this.mediaSourceFactory = mediaSourceFactory;
       return this;
     }
 
@@ -478,13 +457,6 @@ public final class Transformer {
       if (transformationRequest.videoMimeType != null) {
         checkSampleMimeType(transformationRequest.videoMimeType);
       }
-      if (mediaSourceFactory == null) {
-        DefaultExtractorsFactory defaultExtractorsFactory = new DefaultExtractorsFactory();
-        if (transformationRequest.flattenForSlowMotion) {
-          defaultExtractorsFactory.setMp4ExtractorFlags(Mp4Extractor.FLAG_READ_SEF_DATA);
-        }
-        mediaSourceFactory = new DefaultMediaSourceFactory(context, defaultExtractorsFactory);
-      }
       return new Transformer(
           context,
           transformationRequest,
@@ -494,7 +466,6 @@ public final class Transformer {
           removeVideo,
           forceSilentAudio,
           listeners,
-          mediaSourceFactory,
           assetLoaderFactory,
           decoderFactory,
           encoderFactory,
@@ -620,7 +591,6 @@ public final class Transformer {
   private final boolean removeVideo;
   private final boolean forceSilentAudio;
   private final ListenerSet<Transformer.Listener> listeners;
-  private final MediaSource.Factory mediaSourceFactory;
   private final AssetLoader.Factory assetLoaderFactory;
   private final FrameProcessor.Factory frameProcessorFactory;
   private final Muxer.Factory muxerFactory;
@@ -639,7 +609,6 @@ public final class Transformer {
       boolean removeVideo,
       boolean forceSilentAudio,
       ListenerSet<Listener> listeners,
-      MediaSource.Factory mediaSourceFactory,
       AssetLoader.Factory assetLoaderFactory,
       Codec.DecoderFactory decoderFactory,
       Codec.EncoderFactory encoderFactory,
@@ -661,7 +630,6 @@ public final class Transformer {
     this.removeVideo = removeVideo;
     this.forceSilentAudio = forceSilentAudio;
     this.listeners = listeners;
-    this.mediaSourceFactory = mediaSourceFactory;
     this.assetLoaderFactory = assetLoaderFactory;
     this.decoderFactory = decoderFactory;
     this.encoderFactory = encoderFactory;
@@ -804,7 +772,6 @@ public final class Transformer {
             removeAudio,
             removeVideo,
             forceSilentAudio,
-            mediaSourceFactory,
             assetLoaderFactory,
             decoderFactory,
             encoderFactory,